Refactor dashboard components into folders
