+++ src/components/dashboard/OrderHistoryCard/OrderHistoryCard.css
@@ -1,0 +1,63 @@
+.order-chart {
+  position: relative;
+}
+
+.order-chart__canvas {
+  width: 100%;
+  height: 260px;
+}
+
+.order-chart__grid {
+  stroke: #edf1f9;
+  stroke-width: 1;
+}
+
+.order-chart__area {
+  stroke: none;
+}
+
+.order-chart__line {
+  stroke: #6759ff;
+  stroke-width: 4;
+  stroke-linecap: round;
+  stroke-linejoin: round;
+}
+
+.order-chart__dot {
+  fill: #ffffff;
+  stroke: #6759ff;
+  stroke-width: 3;
+}
+
+.order-chart__dot--active {
+  fill: #6759ff;
+}
+
+.order-chart__tooltip {
+  position: absolute;
+  transform: translate(-50%, -120%);
+  background: #111827;
+  color: #ffffff;
+  padding: 10px 14px;
+  border-radius: 12px;
+  text-align: center;
+  box-shadow: 0 12px 24px rgba(15, 23, 42, 0.2);
+}
+
+.order-chart__tooltip-value {
+  display: block;
+  font-weight: 600;
+}
+
+.order-chart__tooltip-label {
+  font-size: 0.75rem;
+  color: rgba(255, 255, 255, 0.7);
+}
+
+.order-chart__axis {
+  display: grid;
+  grid-template-columns: repeat(7, 1fr);
+  font-size: 0.85rem;
+  color: #9ca3af;
+  text-align: center;
+}
