+++ src/components/dashboard/SummaryCards/SummaryCards.css
@@ -1,0 +1,70 @@
+.dashboard__cards {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
+  gap: 20px;
+}
+
+.summary-card {
+  background: #ffffff;
+  border-radius: 24px;
+  padding: 24px;
+  display: flex;
+  gap: 16px;
+  align-items: center;
+  box-shadow: 0 16px 30px rgba(15, 23, 42, 0.04);
+}
+
+.summary-card__badge {
+  width: 56px;
+  height: 56px;
+  border-radius: 20px;
+  flex-shrink: 0;
+}
+
+.summary-card__content h3 {
+  margin: 0;
+  font-size: 0.95rem;
+  color: #6b7280;
+}
+
+.summary-card__value {
+  margin: 10px 0 4px;
+  font-size: 1.6rem;
+  font-weight: 600;
+  color: #111827;
+}
+
+.summary-card__change {
+  display: block;
+  font-size: 0.9rem;
+  color: #34d399;
+  font-weight: 600;
+}
+
+.summary-card__hint {
+  display: block;
+  font-size: 0.82rem;
+  color: #9ca3af;
+}
+
+.summary-card--peach .summary-card__badge {
+  background: linear-gradient(135deg, rgba(249, 115, 22, 0.18), rgba(249, 115, 22, 0.36));
+}
+
+.summary-card--lavender .summary-card__badge {
+  background: linear-gradient(135deg, rgba(129, 140, 248, 0.18), rgba(99, 102, 241, 0.32));
+}
+
+.summary-card--mint .summary-card__badge {
+  background: linear-gradient(135deg, rgba(16, 185, 129, 0.18), rgba(16, 185, 129, 0.32));
+}
+
+.summary-card--sky .summary-card__badge {
+  background: linear-gradient(135deg, rgba(56, 189, 248, 0.18), rgba(37, 99, 235, 0.32));
+}
+
+@media (max-width: 768px) {
+  .dashboard__cards {
+    grid-template-columns: repeat(auto-fit, minmax(180px, 1fr));
+  }
+}
